+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.0] - 2026-03-??
+## [1.4.0] - 2026-03-07 :copilot:
 
 - Add OAS 3.1 support, cross-version warnings, and fix nullable spacing, by @dcode.
 - Fix MARKDOWN style table separators to use minimum 3 hyphens (issue #39), reported by @michael-nok.